+++ src/models/general_structs/metadata.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 use crate::{asset::Asset, location::Location};
 
 #[derive(Clone, CandidType, Serialize, Deserialize, Debug, Default)]
-pub struct GroupMetadata {
+pub struct Metadata {
     pub name: String,
     pub description: String,
     pub website: String,