+++ src/models/general_structs/references.rs
@@ -1,10 +1,8 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 
-use crate::privacy::Privacy;
-
 #[derive(Clone, CandidType, Serialize, Deserialize, Debug, Default)]
-pub struct GroupPrivacy {
-    pub privacy: Privacy,
-    pub privacy_gated_type_amount: Option<u64>,
+pub struct References {
+    pub notification_id: Option<u64>,
+    pub tags: Vec<u32>,
 }