+++ src/models/group_structs/mod.rs
@@ -1,4 +1,0 @@
-pub mod group_members;
-pub mod group_metadata;
-pub mod group_privacy;
-pub mod group_references;