+++ src/models/profile_structs/mod.rs
@@ -1,0 +1,3 @@
+pub mod profile_documents;
+pub mod profile_metadata;
+pub mod profile_references;