+++ src/models/profile_with_refs.rs
@@ -16,82 +16,72 @@ use crate::{
 
 use super::{
     document_details::DocumentDetails,
-    profile_privacy::ProfilePrivacy,
+    general_structs::privacy::Privacy,
+    profile_structs::{
+        profile_documents::ProfileDocuments, profile_metadata::ProfileMetadata,
+        profile_references::ProfileReferences,
+    },
     subject::Subject,
-    wallet::{Wallet, WalletResponse},
+    wallet::WalletResponse,
 };
 
 impl_storable_for!(ProfileWithRefs);
 
 #[derive(Clone, Default, Debug, CandidType, Deserialize, Serialize)]
 pub struct ProfileWithRefs {
-    pub username: String,
-    pub display_name: String,
+    pub metadata: ProfileMetadata,
+    pub documents: ProfileDocuments,
     pub application_role: ApplicationRole,
-    pub first_name: String,
-    pub last_name: String,
-    pub privacy: ProfilePrivacy,
-    pub about: String,
-    pub email: String,
-    pub date_of_birth: u64,
-    pub city: String,
-    pub state_or_province: String,
-    pub country: String,
-    pub profile_image: Asset,
-    pub banner_image: Asset,
-    pub skills: Vec<u32>,
-    pub interests: Vec<u32>,
-    pub causes: Vec<u32>,
-    pub website: String,
-    pub code_of_conduct: Option<DocumentDetails>,
-    pub privacy_policy: Option<DocumentDetails>,
-    pub terms_of_service: Option<DocumentDetails>,
-    pub wallets: HashMap<Principal, Wallet>,
-    pub starred: Vec<Subject>,
-    pub pinned: Vec<Subject>,
-    pub relations: HashMap<Principal, String>,
-    pub extra: String,
+    pub privacy: Privacy,
+    pub references: ProfileReferences,
+    pub extra: Option<String>,
     pub notification_id: Option<u64>,
-    pub groups: Vec<u64>,
-    pub events: Vec<u64>,
     pub updated_on: u64,
     pub created_on: u64,
+}
+
+impl From<PostProfile> for ProfileWithRefs {
+    fn from(profile: PostProfile) -> Self {
+        Self {
+            metadata: ProfileMetadata::from(profile),
+            documents: ProfileDocuments::default(),
+            application_role: ApplicationRole::default(),
+            privacy: Privacy::default(),
+            references: ProfileReferences::default(),
+            extra: None,
+            notification_id: None,
+            updated_on: time(),
+            created_on: time(),
+        }
+    }
 }
 
 impl ProfileWithRefs {
     pub fn update(self, profile: UpdateProfile) -> Self {
         Self {
-            username: self.username,
-            display_name: profile.display_name,
-            application_role: self.application_role,
-            first_name: profile.first_name,
-            last_name: profile.last_name,
+            metadata: ProfileMetadata {
+                display_name: profile.display_name,
+                first_name: profile.first_name,
+                last_name: profile.last_name,
+                about: profile.about,
+                email: profile.email.unwrap_or_default(),
+                date_of_birth: profile.date_of_birth,
+                city: profile.city,
+                state_or_province: profile.state_or_province,
+                country: profile.country,
+                profile_image: profile.profile_image,
+                banner_image: profile.banner_image,
+                website: profile.website,
+                username: self.metadata.username,
+            },
+            documents: self.documents,
             privacy: profile.privacy,
-            about: profile.about,
-            email: profile.email.unwrap_or("".to_string()),
-            date_of_birth: profile.date_of_birth,
-            city: profile.city,
-            state_or_province: profile.state_or_province,
-            country: profile.country,
-            profile_image: profile.profile_image,
-            banner_image: profile.banner_image,
-            skills: profile.skills,
-            interests: profile.interests,
-            causes: profile.causes,
-            website: profile.website,
-            wallets: self.wallets,
-            starred: self.starred,
-            pinned: self.pinned,
-            relations: self.relations,
-            code_of_conduct: self.code_of_conduct,
             extra: profile.extra,
             updated_on: time(),
-            notification_id: self.notification_id,
             created_on: self.created_on,
-            privacy_policy: self.privacy_policy,
-            terms_of_service: self.terms_of_service,
-            groups: vec![],
-            events: vec![],
+            application_role: self.application_role,
+            references: self.references,
+            notification_id: self.notification_id,
         }
     }
 
@@ -104,89 +94,51 @@ impl ProfileWithRefs {
     }
 
     pub fn is_starred(&self, subject: &Subject) -> bool {
-        self.starred.contains(subject)
+        self.references.starred.contains(subject)
     }
 
     pub fn is_pinned(&self, subject: &Subject) -> bool {
-        self.pinned.contains(subject)
+        self.references.pinned.contains(subject)
     }
 
     pub fn remove_pinned(&mut self, subject: &Subject) {
-        self.pinned.retain(|s| s != subject);
+        self.references.pinned.retain(|s| s != subject);
     }
 
     pub fn remove_starred(&mut self, subject: &Subject) {
-        self.starred.retain(|s| s != subject);
+        self.references.starred.retain(|s| s != subject);
     }
 
     pub fn add_group(&mut self, group_id: u64) {
-        self.groups.push(group_id);
+        self.references.groups.push(group_id);
     }
 
     pub fn is_group_member(&self, group_id: u64) -> bool {
-        self.groups.contains(&group_id)
+        self.references.groups.contains(&group_id)
     }
 
     pub fn get_group_ids(&self) -> Vec<u64> {
-        self.groups.clone()
+        self.references.groups.clone()
     }
 
     pub fn remove_group(&mut self, group_id: u64) {
-        self.groups.retain(|id| id != &group_id);
+        self.references.groups.retain(|id| id != &group_id);
     }
 
     pub fn add_event(&mut self, event_id: u64) {
-        self.events.push(event_id);
+        self.references.events.push(event_id);
     }
 
     pub fn is_event_attendee(&self, event_id: u64) -> bool {
-        self.events.contains(&event_id)
+        self.references.events.contains(&event_id)
     }
 
     pub fn get_event_ids(&self) -> Vec<u64> {
-        self.events.clone()
+        self.references.events.clone()
     }
 
     pub fn remove_event(&mut self, event_id: u64) {
-        self.events.retain(|id| id != &event_id);
-    }
-}
-
-impl From<PostProfile> for ProfileWithRefs {
-    fn from(profile: PostProfile) -> Self {
-        Self {
-            username: profile.username,
-            display_name: profile.display_name,
-            application_role: ApplicationRole::default(),
-            first_name: profile.first_name,
-            last_name: profile.last_name,
-            privacy: profile.privacy,
-            about: "".to_string(),
-            email: "".to_string(),
-            date_of_birth: 0,
-            city: "".to_string(),
-            state_or_province: "".to_string(),
-            country: "".to_string(),
-            profile_image: Asset::None,
-            banner_image: Asset::None,
-            skills: vec![],
-            interests: vec![],
-            causes: vec![],
-            website: "".to_string(),
-            wallets: HashMap::new(),
-            starred: Vec::new(),
-            pinned: Vec::new(),
-            relations: HashMap::new(),
-            code_of_conduct: None,
-            extra: profile.extra,
-            updated_on: time(),
-            created_on: time(),
-            notification_id: None,
-            privacy_policy: None,
-            terms_of_service: None,
-            groups: vec![],
-            events: vec![],
-        }
+        self.references.events.retain(|id| id != &event_id);
     }
 }
 
@@ -198,7 +150,7 @@ pub struct PostProfile {
     pub display_name: String,
     pub first_name: String,
     pub last_name: String,
-    pub privacy: ProfilePrivacy,
+    pub privacy: Privacy,
     pub extra: String,
 }
 
@@ -207,7 +159,7 @@ pub struct UpdateProfile {
     pub display_name: String,
     pub first_name: String,
     pub last_name: String,
-    pub privacy: ProfilePrivacy,
+    pub privacy: Privacy,
     pub about: String,
     pub email: Option<String>,
     pub date_of_birth: u64,
@@ -220,7 +172,7 @@ pub struct UpdateProfile {
     pub interests: Vec<u32>,
     pub causes: Vec<u32>,
     pub website: String,
-    pub extra: String,
+    pub extra: Option<String>,
 }
 
 #[derive(Clone, Debug, CandidType, Deserialize, Serialize)]
@@ -231,7 +183,7 @@ pub struct ProfileResponse {
     pub application_role: ApplicationRole,
     pub first_name: String,
     pub last_name: String,
-    pub privacy: ProfilePrivacy,
+    pub privacy: Privacy,
     pub about: String,
     pub email: String,
     pub date_of_birth: u64,
@@ -250,7 +202,7 @@ pub struct ProfileResponse {
     pub pinned: Vec<Subject>,
     pub starred: Vec<Subject>,
     pub wallets: Vec<WalletResponse>,
-    pub extra: String,
+    pub extra: Option<String>,
     pub updated_on: u64,
     pub created_on: u64,
 }
@@ -258,41 +210,42 @@ pub struct ProfileResponse {
 impl ProfileResponse {
     pub fn new(principal: Principal, profile: ProfileWithRefs) -> Self {
         let wallets = profile
+            .references
             .wallets
             .into_iter()
-            .map(|(principal, wallet)| WalletResponse {
+            .map(|(address, wallet)| WalletResponse {
                 provider: wallet.provider,
-                principal,
+                address,
                 is_primary: wallet.is_primary,
             })
             .collect();
 
         Self {
             principal,
-            username: profile.username,
-            display_name: profile.display_name,
-            about: profile.about,
-            city: profile.city,
-            country: profile.country,
-            website: profile.website,
-            skills: profile.skills,
-            interests: profile.interests,
-            causes: profile.causes,
-            email: profile.email,
+            username: profile.metadata.username,
+            display_name: profile.metadata.display_name,
+            about: profile.metadata.about,
+            city: profile.metadata.city,
+            country: profile.metadata.country,
+            website: profile.metadata.website,
+            skills: profile.references.skills,
+            interests: profile.references.interests,
+            causes: profile.references.causes,
+            email: profile.metadata.email,
             application_role: profile.application_role,
-            first_name: profile.first_name,
-            last_name: profile.last_name,
+            first_name: profile.metadata.first_name,
+            last_name: profile.metadata.last_name,
             privacy: profile.privacy,
-            date_of_birth: profile.date_of_birth,
-            state_or_province: profile.state_or_province,
-            profile_image: profile.profile_image,
-            banner_image: profile.banner_image,
-            code_of_conduct: profile.code_of_conduct,
-            privacy_policy: profile.privacy_policy,
-            terms_of_service: profile.terms_of_service,
+            date_of_birth: profile.metadata.date_of_birth,
+            state_or_province: profile.metadata.state_or_province,
+            profile_image: profile.metadata.profile_image,
+            banner_image: profile.metadata.banner_image,
+            code_of_conduct: profile.documents.code_of_conduct,
+            privacy_policy: profile.documents.privacy_policy,
+            terms_of_service: profile.documents.terms_of_service,
             wallets,
-            pinned: profile.pinned,
-            starred: profile.starred,
+            pinned: profile.references.pinned,
+            starred: profile.references.starred,
             extra: profile.extra,
             updated_on: profile.updated_on,
             created_on: profile.created_on,
@@ -307,41 +260,42 @@ impl ProfileResponse {
 impl From<ProfileEntry> for ProfileResponse {
     fn from((principal, profile): ProfileEntry) -> Self {
         let wallets = profile
+            .references
             .wallets
             .into_iter()
-            .map(|(principal, wallet)| WalletResponse {
+            .map(|(address, wallet)| WalletResponse {
                 provider: wallet.provider,
-                principal,
+                address,
                 is_primary: wallet.is_primary,
             })
             .collect();
 
         Self {
             principal,
-            username: profile.username,
-            display_name: profile.display_name,
-            about: profile.about,
-            city: profile.city,
-            country: profile.country,
-            website: profile.website,
-            skills: profile.skills,
-            interests: profile.interests,
-            causes: profile.causes,
-            email: profile.email,
+            username: profile.metadata.username,
+            display_name: profile.metadata.display_name,
+            about: profile.metadata.about,
+            city: profile.metadata.city,
+            country: profile.metadata.country,
+            website: profile.metadata.website,
+            skills: profile.references.skills,
+            interests: profile.references.interests,
+            causes: profile.references.causes,
+            email: profile.metadata.email,
             application_role: profile.application_role,
-            first_name: profile.first_name,
-            last_name: profile.last_name,
+            first_name: profile.metadata.first_name,
+            last_name: profile.metadata.last_name,
             privacy: profile.privacy,
-            date_of_birth: profile.date_of_birth,
-            state_or_province: profile.state_or_province,
-            profile_image: profile.profile_image,
-            banner_image: profile.banner_image,
-            code_of_conduct: profile.code_of_conduct,
-            privacy_policy: profile.privacy_policy,
-            terms_of_service: profile.terms_of_service,
+            date_of_birth: profile.metadata.date_of_birth,
+            state_or_province: profile.metadata.state_or_province,
+            profile_image: profile.metadata.profile_image,
+            banner_image: profile.metadata.banner_image,
+            code_of_conduct: profile.documents.code_of_conduct,
+            privacy_policy: profile.documents.privacy_policy,
+            terms_of_service: profile.documents.terms_of_service,
             wallets,
-            pinned: profile.pinned,
-            starred: profile.starred,
+            pinned: profile.references.pinned,
+            starred: profile.references.starred,
             extra: profile.extra,
             updated_on: profile.updated_on,
             created_on: profile.created_on,
@@ -401,20 +355,20 @@ impl Filter<Principal, ProfileWithRefs> for ProfileFilter {
     fn matches(&self, _key: &Principal, value: &ProfileWithRefs) -> bool {
         use ProfileFilter::*;
         match self {
-            Username(username) => eq_str(&value.username, username),
-            DisplayName(display_name) => eq_str(&value.display_name, display_name),
-            FirstName(first_name) => eq_str(&value.first_name, first_name),
-            LastName(last_name) => eq_str(&value.last_name, last_name),
-            Email(email) => eq_str(&value.email, email),
-            City(city) => eq_str(&value.city, city),
+            Username(username) => eq_str(&value.metadata.username, username),
+            DisplayName(display_name) => eq_str(&value.metadata.display_name, display_name),
+            FirstName(first_name) => eq_str(&value.metadata.first_name, first_name),
+            LastName(last_name) => eq_str(&value.metadata.last_name, last_name),
+            Email(email) => eq_str(&value.metadata.email, email),
+            City(city) => eq_str(&value.metadata.city, city),
             StateOrProvince(state_or_province) => {
-                eq_str(&value.state_or_province, state_or_province)
+                eq_str(&value.metadata.state_or_province, state_or_province)
             }
-            Country(country) => eq_str(&value.country, country),
+            Country(country) => eq_str(&value.metadata.country, country),
             UpdatedOn(date_range) => date_range.is_within(value.updated_on),
-            Skill(skill) => value.skills.contains(skill),
-            Interest(interest) => value.interests.contains(interest),
-            Cause(cause) => value.causes.contains(cause),
+            Skill(skill) => value.references.skills.contains(skill),
+            Interest(interest) => value.references.interests.contains(interest),
+            Cause(cause) => value.references.causes.contains(cause),
             CreatedOn(date_range) => date_range.is_within(value.created_on),
         }
     }